// src/lib/api/generated/types.ts
export interface components {
    schemas: {
// ...
        EventResponseDTO: {
            id?: string;
            title?: string;
            description?: string;
            ticketUrl?: string;
            venue?: string;
            city?: string;
            /** @enum {string} */
            region?: "KYIV_CITY" | "SEVASTOPOL" | "CRIMEA" | "LVIV" | "VOLYN" | "RIVNE" | "TERNOPIL" | "IVANO_FRANKIVSK" | "ZAKARPATTIA" | "CHERNIVTSI" | "KHMELNYTSKYI" | "ZHYTOMYR" | "KYIV" | "CHERNIHIV" | "SUMY" | "VINNYTSIA" | "CHERKASY" | "POLTAVA" | "ODESA" | "MYKOLAIV" | "KHERSON" | "KIROVOHRAD" | "DNIPROPETROVSK" | "KHARKIV" | "ZAPORIZHZHIA" | "DONETSK" | "LUHANSK";
            /** Format: double */
            latitude?: number;
            /** Format: double */
            longitude?: number;
            geohash?: string;
            /** Format: date-time */
            dateTime?: string;
            /** @enum {string} */
            status?: "PENDING" | "APPROVED" | "REJECTED";
            eventImageKey?: string;
            artistName?: string;
            slug?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}